libcros: Fix SConstruct and correct build warnings.
The SConstruct was setting -Wall and -Werror but then
clearing them by copying in CCFLAGS from the ebuild.  As a
result, a few bugs (uninitialized variables, missing return
values, etc.) weren't causing build failures.  This fixes
the SConstruct and the bugs.


BUG=none
TEST=built it

Review URL: http://codereview.chromium.org/3346008

+++ chromeos_input_method_ui.cc
@@ -271,7 +271,7 @@ gboolean ibus_chromeos_panel_service_update_lookup_table(
       }
 
       // Check that the attribute indicates annotation.
-      if (IsAnnotation(attr) && j + 1 == length) {
+      if (IsAnnotation(attr) && j + 1 == static_cast<int>(length)) {
         const std::wstring candidate_word =
             UTF8ToWide(text->text).substr(0, attr->start_index);
         lookup_table.candidates.push_back(WideToUTF8(candidate_word));
